Refactor proposal forms into subpackage
