+++ src/proposals/forms/__init__.py
@@ -1,0 +1,2 @@
+from .proposals import *    # noqa
+from .speakers import *     # noqa
