+++ src/proposals/forms/speakers.py
@@ -2,50 +2,16 @@ from django import forms
 from django.contrib.auth import get_user_model
 from django.utils.translation import ugettext_lazy as _
 
-from core.widgets import SimpleMDEWidget
+from proposals.models import AdditionalSpeaker
 
-from .models import AdditionalSpeaker, TalkProposal, TutorialProposal
+from .mixins import RequestUserSpeakerValidationMixin
 
 
 User = get_user_model()
 
 
-class OwnedByRequestUserMixin:
-    """Mixin providing ``self._request`` and auth validation on cleaning.
-    """
-    error_messages = {
-        'no_request': _(
-            '{model_name_cap} creation requires a request object.'
-        ),
-        'bad_speaker': _(
-            'Only authenticated user with complete speaker profile may '
-            'create a {model_name}.'
-        ),
-    }
-
-    def __init__(self, request=None, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._request = request
-
-    def clean(self):
-        """Validate user for saving later.
-        """
-        if self._request is None:
-            raise forms.ValidationError(self.get_error_message('no_request'))
-        user = self._request.user
-        if user.is_anonymous() or not user.is_valid_speaker():
-            raise forms.ValidationError(self.get_error_message('bad_speaker'))
-        return self.cleaned_data
-
-    def get_error_message(self, key):
-        model_name = self._meta.model._meta.verbose_name
-        return self.error_messages[key].format(
-            model_name=model_name,
-            model_name_cap=model_name.capitalize(),
-        )
-
-
-class AdditionalSpeakerCreateForm(OwnedByRequestUserMixin, forms.ModelForm):
+class AdditionalSpeakerCreateForm(
+        RequestUserSpeakerValidationMixin, forms.ModelForm):
     """Form used to add an additional speaker to a proposal.
     """
     email = forms.EmailField(
@@ -56,7 +22,7 @@ class AdditionalSpeakerCreateForm(OwnedByRequestUserMixin, forms.ModelForm):
         ),
     )
 
-    error_messages = OwnedByRequestUserMixin.error_messages.copy()
+    error_messages = RequestUserSpeakerValidationMixin.error_messages.copy()
     error_messages.update({
         'bad_speaker': _(
             'Only authenticated user with complete speaker profile may '
@@ -173,98 +139,3 @@ class AdditionalSpeakerSetStatusForm(AdditionalSpeakerUpdateForm):
     class Meta:
         model = AdditionalSpeaker
         fields = ['status']
-
-
-class ProposalCreateForm(OwnedByRequestUserMixin, forms.ModelForm):
-    """Form used to create a proposal.
-
-    Fields in this form is intentionally reduced to allow people to submit
-    a proposal very quickly, and fill in the details later.
-    """
-    def save(self, commit=True):
-        """Fill user field on save.
-        """
-        proposal = super().save(commit=False)
-        proposal.submitter = self._request.user
-        if commit:
-            proposal.save()
-        return proposal
-
-
-class TalkProposalCreateForm(ProposalCreateForm):
-    class Meta:
-        model = TalkProposal
-        fields = [
-            'title', 'category', 'duration', 'language',
-            'python_level', 'recording_policy',
-        ]
-
-
-class TutorialProposalCreateForm(ProposalCreateForm):
-    class Meta:
-        model = TutorialProposal
-        fields = [
-            'title', 'category', 'duration', 'language',
-            'python_level', 'recording_policy',
-        ]
-
-
-class TalkProposalUpdateForm(forms.ModelForm):
-    """Form used to update a talk proposal.
-
-    This is the complete editing form for proposal. It should contain all
-    user-editable fields.
-    """
-    class Meta:
-        model = TalkProposal
-        fields = [
-            'title', 'category', 'duration', 'language', 'target_audience',
-            'abstract', 'python_level', 'objective', 'detailed_description',
-            'outline', 'supplementary', 'recording_policy', 'slide_link',
-        ]
-        widgets = {
-            'detailed_description': SimpleMDEWidget(),
-            'outline': SimpleMDEWidget(),
-            'supplementary': SimpleMDEWidget(),
-        }
-
-
-class TutorialProposalUpdateForm(forms.ModelForm):
-    """Form used to update a tutorial proposal.
-
-    This is the complete editing form for proposal. It should contain all
-    user-editable fields.
-    """
-    class Meta:
-        model = TutorialProposal
-        fields = [
-            'title', 'category', 'duration', 'language', 'target_audience',
-            'abstract', 'python_level', 'objective', 'detailed_description',
-            'outline', 'supplementary', 'recording_policy', 'slide_link',
-        ]
-        widgets = {
-            'detailed_description': SimpleMDEWidget(),
-            'outline': SimpleMDEWidget(),
-            'supplementary': SimpleMDEWidget(),
-        }
-
-
-class ProposalCancelForm(forms.ModelForm):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        if not self.instance:
-            raise ValueError(
-                'Proposal cancel form must be initialized with an instance.'
-            )
-
-
-class TalkProposalCancelForm(ProposalCancelForm):
-    class Meta:
-        model = TalkProposal
-        fields = ['cancelled']
-
-
-class TutorialProposalCancelForm(ProposalCancelForm):
-    class Meta:
-        model = TutorialProposal
-        fields = ['cancelled']
